# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AVIS — Agentic Visual Incident Investigation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all uploaded videos
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: uploadVideo
      tags: [videos]
      summary: Upload a new video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /videos/{id}/process:
    post:
      operationId: processVideo
      tags: [videos]
      summary: Trigger processing pipeline for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Video:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        cameraName:
          type: string
        status:
          type: string
        uploadTime:
          type: string
          format: date-time
        processedAt:
          type: ["string", "null"]
          format: date-time
        durationSeconds:
          type: ["integer", "null"]
        frameCount:
          type: ["integer", "null"]
        thumbnailUrl:
          type: ["string", "null"]
      required:
        - id
        - fileName
        - cameraName
        - status
        - uploadTime
    VideoInput:
      type: object
      properties:
        fileName:
          type: string
        cameraName:
          type: string
        durationSeconds:
          type: integer
      required:
        - fileName
        - cameraName
...