# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AVIS — Agentic Visual Incident Investigation System API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  fileName: string;
  cameraName: string;
  status: string;
  uploadTime: Date;
  /** @nullable */
  processedAt?: Date | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  frameCount?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
}